fix: refine floatbar local cost handling
