+++ apps/desktop-tauri/src/i18n/keys.ts
@@ -456,6 +456,9 @@ export const ALL_LOCALE_KEYS = [
   "PanelReserveLastsUntilReset",
   "PanelReserveRunsOutInDaysHours",
   "PanelReserveRunsOutInHours",
+  "FloatBarThirtyDayShort",
+  "FloatBarNoProviders",
+  "FloatBarRemainingSuffix",
   "BannerCheckingForUpdates",
   "BannerUpdateAvailablePrefix",
   "BannerDownloadButton",
