+++ apps/desktop-tauri/src/surfaces/TrayPanel.tsx
@@ -1,5 +1,4 @@
 import { Fragment, useCallback, useEffect, useMemo, useRef, useState, type CSSProperties } from "react";
-import { listen } from "@tauri-apps/api/event";
 import { getCurrentWindow } from "@tauri-apps/api/window";
 import type { BootstrapState, ProviderUsageSnapshot } from "../types/bridge";
 import {
@@ -10,7 +9,6 @@ import {
   openSettingsWindow,
   quitApp as quitApplication,
   reorderProviders,
-  refreshProvidersIfStale,
   setFlyoutSize,
   setSurfaceMode,
   updateSettings,
@@ -104,21 +102,6 @@ export default function TrayPanel({ state }: { state: BootstrapState }) {
   const { updateState, checkNow, download, apply, dismiss, openRelease } =
     useUpdateState();
 
-  useEffect(() => {
-    let active = true;
-    const unlisten = listen("flyout-opened", () => {
-      if (!active) return;
-      if (settings.refreshAllProvidersOnMenuOpen) {
-        refresh();
-      } else {
-        void refreshProvidersIfStale();
-      }
-    });
-    return () => {
-      active = false;
-      unlisten.then((fn) => fn());
-    };
-  }, [refresh, settings.refreshAllProvidersOnMenuOpen]);
   const { t } = useLocale();
   const surfaceTarget = useSurfaceTarget("trayPanel");
