+++ rust/src/locale/en-US.ftl
@@ -449,6 +449,9 @@ PanelReserveSuffix = in reserve
 PanelReserveLastsUntilReset = Lasts until reset
 PanelReserveRunsOutInDaysHours = Runs out in { "{}" }d { "{}" }h
 PanelReserveRunsOutInHours = Runs out in { "{}" }h
+FloatBarThirtyDayShort = 30d
+FloatBarNoProviders = No providers
+FloatBarRemainingSuffix = remaining
 BannerCheckingForUpdates = Checking for updates…
 BannerUpdateAvailablePrefix = Update
 BannerDownloadButton = Download
