+++ rust/src/locale/es-MX.ftl
@@ -449,6 +449,9 @@ PanelReserveSuffix = en reserva
 PanelReserveLastsUntilReset = Dura hasta el reinicio
 PanelReserveRunsOutInDaysHours = Se agota en { "{}" }d { "{}" }h
 PanelReserveRunsOutInHours = Se agota en { "{}" }h
+FloatBarThirtyDayShort = 30d
+FloatBarNoProviders = Sin proveedores
+FloatBarRemainingSuffix = restante
 BannerCheckingForUpdates = Buscando actualizaciones…
 BannerUpdateAvailablePrefix = Actualización
 BannerDownloadButton = Descargar
