+++ rust/src/locale/ja-JP.ftl
@@ -449,6 +449,9 @@ PanelReserveSuffix = 予備
 PanelReserveLastsUntilReset = リセットまで持ちます
 PanelReserveRunsOutInDaysHours = 残り { "{}" }日 { "{}" }時間
 PanelReserveRunsOutInHours = 残り { "{}" }時間
+FloatBarThirtyDayShort = 30日
+FloatBarNoProviders = プロバイダーなし
+FloatBarRemainingSuffix = 残り
 BannerCheckingForUpdates = Checking for updates…
 BannerUpdateAvailablePrefix = Update
 BannerDownloadButton = Download
