+++ rust/src/locale/ko-KR.ftl
@@ -449,6 +449,9 @@ PanelReserveSuffix = 예비
 PanelReserveLastsUntilReset = 리셋까지 지속
 PanelReserveRunsOutInDaysHours = { "{}" }일 { "{}" }시간 후 소진
 PanelReserveRunsOutInHours = { "{}" }시간 후 소진
+FloatBarThirtyDayShort = 30일
+FloatBarNoProviders = 공급자 없음
+FloatBarRemainingSuffix = 남음
 BannerCheckingForUpdates = 업데이트 확인 중…
 BannerUpdateAvailablePrefix = 업데이트
 BannerDownloadButton = 다운로드
