+++ rust/src/locale/zh-CN.ftl
@@ -449,6 +449,9 @@ PanelReserveSuffix = 储备
 PanelReserveLastsUntilReset = 持续到重置
 PanelReserveRunsOutInDaysHours = { "{}" }天{ "{}" }小时后耗尽
 PanelReserveRunsOutInHours = { "{}" }小时后耗尽
+FloatBarThirtyDayShort = 30天
+FloatBarNoProviders = 无提供商
+FloatBarRemainingSuffix = 剩余
 BannerCheckingForUpdates = 正在检查更新…
 BannerUpdateAvailablePrefix = 更新
 BannerDownloadButton = 下载
